requirements.txt
aws-cdk.core>=1.131.0
aws-cdk.assertions>=1.131.0
aws-cdk.aws_s3>=1.31.0
aws-cdk.aws_lambda>=1.31.0